+++ packages/utilities/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @module-federation/utilities
 
+## 3.1.82
+
+### Patch Changes
+
+- @module-federation/sdk@0.24.1
+
+## 3.1.81
+
+### Patch Changes
+
+- @module-federation/sdk@0.24.0
+
 ## 3.1.80
 
 ### Patch Changes